Add support for compute API v2.70 - os-volume_attachments
This adds the compute API response schema validation for the
2.70 microversion and the following APIs:

- GET /servers/{server_id}/os-volume_attachments (list)
- GET /servers/{server_id}/os-volume_attachments/{volume_id} (show)
- POST /servers/{server_id}/os-volume_attachments (attach)

This will be used in an upcoming change for testing root volume
attach/detach of a shelved offloaded server.

Since the response schema for these have not changed since 2.1 we
have to backfill all of the other servers API response schema
files going back to 2.1.

Change-Id: I2f36a51fa47b28df1afada5396dec3ce7dc43d78