Use credentials domain instead of creating new one
test_list_projects tests create a new domain and try to test the
list projects operation based on new domain-id. Keystone is
implementing the domain roles functionality for projects resource
- https://review.openstack.org/#/c/624218/

Which means listing the projects with authorized domain only.

This commit modify the tests to use the same domain as requester and
create project with that domain only not new one.

Also add the debug log.

Change-Id: I401815b4a0d3f7ad90801d5580897be870d33013