Make type decorator compatible with unicode strings
When importing unicode_literals from __future__ in tempest plugins as in
watcher, type is set as unicode instead of string when running in
python2, so the type is not properly detected. This is making that only
3 tests are detected as smoke in watcher tests plugins with python2 (other
may be affected) instead of 52 which are discovered as smoke with python3.

This patch uses six to make this compatible with both string and unicode
in python2.

Change-Id: Ibba9d8217f543fae0c232757ddb18c4fdf957aff
Co-Authored-By: yatin <ykarel@redhat.com>