Set ENABLE_VOLUME_MULTIATTACH=true on tempest-slow subnode localrc
Before change Icb971831c8d4fe5f940d9e7993d53f1c3765e30f in devstack
in Rocky, the ENABLE_VOLUME_MULTIATTACH variable is used to control
whether or not the UCA is enabled on the node and defaults to False.
The tempest-slow job is multi-node and runs on stable branches since
change Ie5492603a12ca4ae56baa744bf650ae4d8036256 and as a result
is configuring the UCA on the subnode but not the controller, so
nova-compute on each node is running with different versions of QEMU
which can cause live migration failures if trying to live migrate
from a newer version of QEMU to an older version.

This fixes the issue by configuring the subnode with the correct
localrc value for ENABLE_VOLUME_MULTIATTACH.

Change-Id: Idea94dde6973805995c1678fcfb6b0e93025492a
Closes-Bug: #1819910