Delete test_reboot_server_soft
The test has been skipped for than 6 years already.

Nova automatically switched to hard reboot if the guest is
not responding, no way to see the difference from the API.

The following alternatives were not agreed on the PTG.
 - Change the acpid config an all supported image before the reboot,
   and log the acpi event in a tempest friendly way
 - ssh the machine before reboot, in this case more likely a
   soft reboot would happen, but we are unable to distinguish it from
   the hard one.

Take into account that the minimum scenario test uses soft reboot
and the nova functional test also covers reboot.

The test hasn't failed for more than 6 years (it's been skipped),
so nothing prevents us from removing it by the usual removal procedure:

 https://docs.openstack.org/tempest/latest/test_removal.html

The test deletion was also announced on ML:
 http://lists.openstack.org/pipermail/openstack-discuss/2020-October/017889.html

Change-Id: I62b48865f5b21e55c28b8ee08ad5786473cc5ddf
Related-Bug: #1014647