Improve router deletion logging in tempest cleanup
Improve logging in tempest cleanup when a router and
its interfaces are being deleted. The review adds a
try except block for every attempt to remove a port.
By that addition any port error will not be hidden
behind the router error the port is attached to.
It will provide more precise error logging.

Change-Id: I475deec7b29600627f68ff07c5546e55cdab100f