Add APIMicroversionFixture to library interface
APIMicroversionFixture is used to set the microversion on
service clients which will take care of reseting the microversion.

This interface has been added to set the microversion on all supported
service clients which use microverison. User can set multiple
service microversion all together.

Using the new interface in compute tests to make sure it is all
working in serial and parallel testing.

Change-Id: I6ed41b02041d38c2c92c58b0d3dd6613d03da3f7