Handle 'path' query parameter for test_novnc
Starting in v1.1.0 of noVNC, it no longer forwards a token from the
URI via a cookie and we must instead specify the token as part of the
'path' query parameter.

This adds handling of the 'path' query parameter so the test_novnc
tests will be able to work with a console access url that has the token
in the 'path' query parameter instead of in a cookie and thus work with
noVNC >= v1.1.0.

Related-Bug: #1822676

Change-Id: Ifc0262f869e4d08d2746275438575980213fe9b2