Improve ipv6 configuration
The change replaces restarting the network by reloading network
profiles. Restart of the network terminates ssh connection, so
if possible, better to avoid it. This approach is also
compatible with RHEL 8 as there's no network service anymore
and therefore can't be restarted.
The patch also brings the configured interface up by nmcli,
which works better if there already was a network profile
connected to the configured interface. nmcli makes it drop the
existing profile and use the one which was created in the test.

Change-Id: I4ea06f28c525f2631a69f759fc96b262d9100f45