Revert "Refactor random naming"
This reverts commit e3e7b2722e6b78216373f392af5bfeba1e2da5ec.

After that change, if the test creates a volume type with a
given name, it's not unique and therefore can hit a 409 conflict
trying to create a volume type with a name that already exists,
which can happen when running scenario tests concurrently since
several encrypted volume tests use the same volume type name "luks".

See Iecf62d411d2da17c4c983915018e3ed9180ecd11 for an
alternative but this is easier. It's not entirely clear
why the original change was made anyway - why do we care
if the tests create long names? They weren't unnecessary
because they were random to avoid conflicts.

Change-Id: I9f5a967514f3c79f861d286f65402e9ec2832cc4
Closes-Bug: #1826953