Remove airship-tempest-plugin from blacklist
This commit removes airship-tempest-plugin entry from the BLACKLIST
since the patch was already merged. And airship/tempest-plugin has a
gate job to verify it, too. There is no reason to store it in the
BLACKLIST anymore.

And this commit also make plugin-sanity-check's basepython python3
because we should use it as a default.

Change-Id: I9c50d327df65fecf8510f6f54f06e9f42da9bea0