Add cyborg-tempest-plugin to required-projects and remove octavia
This commit adds opendev.org/openstack/cyborg-tempest-plugin to
required-projects to pass the plugin-sanity-check job. It's a new
tempest plugin. It doesn't work properly on master yet. So, I put it on
the BLACKLIST.

And this commit also removes octavia because octavia tempest tests were
already migrated to octavia-tempest-plugin, and there is no tempest test
in the octavia repo. The removal patch[0] in octavia was already merged.

[0] https://review.opendev.org/#/c/659516/

Change-Id: Ied13dbf774472e1c36dca4ccca157104a81fbf9d