Change IPv6 ICMP security group rule check
A neutron change, https://review.opendev.org/#/c/453346/
is standardizing the protocol name for IPv6 ICMP in security
group rules to be 'ipv6-icmp', even if 'icmp' or 'icmpv6'
was passed during creation.  Change the IPv6 ICMP test to use
'ipv6-icmp' as the protocol name when the ethertype is IPv6.

Change-Id: I7c0e0caf67ab39743cdd7600dba099aecacd86db
Partial-Bug: #1582500