Publish non-openstack namespace tempest plugins
This commit changes to publish non-openstack namespace tempest plugins
to the tempest registry doc page. We missed some tempest plugins such as
airship's tempest-plugin because of the recent namespace change. And
also, this commit adds two projects[0] to .zuul.yaml because it had
been added already.

This commit republish them on the doc page, and add two projects to the
blacklist to pass the sanity tests.

Change-Id: Ieecb8989ff8959f1b2c0e5f2f6b8ea2c3ebbcaa9