Move tox execution to current docs section
This commit moves the execution explanation of tempest run to the
current(non-legacy) docs section. In the original section of the tox
jobs are "Legacy run method" which discourages to use it. However, we
don't do it yet so far. So, this commit would make it clearer for using
the tox jobs as a tempest runner.

Change-Id: Iec6a956cba6949d0f30b0d0297ed26487e892125