Revert "Skip test_server_connectivity_cold_migration_revert until fixed"
This reverts commit 2a4641c056f3ca0157d1c4ea6cd8e45f85e9213e.

Due to change I9e0cffb889c94713c7f28812918103a5d97cefeb in nova
the bug is likely fixed in the gate so we can unskip it for our
standard gate testing.

Note that this change will depend on a nova patch to fix how
network-vif-plugged event handling works for OVS hybrid type
ports to (1) make sure that change works in a multi-node scenario
and (2) to not unskip this test until that bug is fixed.

Depends-On: https://review.opendev.org/667177/

Change-Id: I3911994392ec2050f5d5b4bbf83de4f329b6e372
Related-Bug: #1788403