Add 'single core approve' policy for plugins sanity jobs failure
We discussed in Denver PTG[1] to make plugins
sanity jobs as voting which will help to mantain
the active plugins and blacklist the faulty plugins.

But when we get faulty plugins this job which will
be voting after https://review.opendev.org/#/c/641188/
can block the Tempest gate. So adding the process of
single core approve policy for mergin the patches which
will fix or remove the faulty plugins to unblock the gate.

[1] L196 https://etherpad.openstack.org/p/qa-train-ptg

Change-Id: Id38e6270ac1bae8ba63761f9bab6f1e5d0a2252b