Skip adding host to aggregate with az if host already in other zone
Running test_aggregate_basic_ops fails if the selected host is
already in an other zone:

"Cannot add host to aggregate 1773. Reason: One or more hosts already
in availability zone(s)"

This is to find an up and enabled host that has not been added to
other zone, and skip test_aggregate_basic_ops if all hosts in
the system are already in other zones.

Change-Id: I2c196ca22c963af976b256016c36f4095ccf8f61