Make bashate E005,E042 error and ignore E006
This commit makes two bashate rules E005 and E042 treat as error and
ignores E006. It is difficult to keep E006 rule because bash scripts
tend to be long naturally. Both E005 and E042 are good to be errors
since it should be safer.

These additions can be seen in devstack[0] too.

E005: file does not begin with #! or have a .sh prefix
E006: check for lines longer than 79 columns
E042: local declaration hides errors

[0] https://opendev.org/openstack/devstack/src/commit/0fc6b2c5a87a1884bf3cff086081b3a253cd9302/tox.ini#L35

Change-Id: I56c017569aa8660648764502cce5adfe2e946a58