tempest cleanup fix return code
It's a follow up fix for tempest cleanup NotImplemented error issue.
Tempest cleanup shouldn't return a non-zero return code for
NotImplemented errors as they are actually not errors, they are just
an outcome of tempest not checking availability of extensions in order
to keep it simple without any complicated logic.

Closes-bug: #1832566

Change-Id: I09e28c245c11dd3dd061df786bd137eceb97be07