Remove references to locking test accounts config doc
The name locking test accounts was an early name in the development of
the pre-provisioned credentials development. It described what it did
used tests accounts and lock files to ensure one test class only was
using an account at a time. However, that name hasn't been really used
in a long time and we've called the feature pre-provisioned credentials.
The configuration guide section for pre-provisioned credentials still
had some out dated references to locking test accounts which were never
updated when the feature was renamed. This commit corrects the oversight
and consistently names the feature in that section of the documentation.

Change-Id: I0f3a7b373c5ca2d84a8e82c1e2c3bcd6a2d833f9