Add rand_name to create port incase of leaks
We have found if people either kill a tempest run
or miss-configure tempest.conf we can get ports
left around.  They are easier to clean up if they
have an explicit name with the default prefix
and the class name of the test that created the
port than a standard port name that is genreated
by the system.

FYI neutron port name is string 255 in the neutron database.

Change-Id: I8d96387c0f8c70e7e4ea004a8419689f1bd30cef