"test_project_get_equals_list" allows extra fields
Keystone is updating the project response to include resource-specific
options. The test `test_project_get_equals_list` is looking at explicit
data being returned and prevents the additional field. Keystone does not
(and currently has not plan) to support microversions. This changes the
test checking the returned fields to ensure that the expected fields is
a subset (inclusive) of the returned fields. This allows for Keystone to
iterate and respond with additional fields for future changes.

Any future fields added become part of the contract and should be added
to the expected "fields" list in the test after the new field response
code lands within keystone.

Related-Bug: #1807751
Required-by: https://review.opendev.org/#/c/678322/
Change-Id: I266d98503066f3a8027effc43a95f9ad9ff12492