Fix use of get_server_ip with fixed networks
If a fixed network is used, we see errors like the following in three
compute tests:

KeyError: 'addresses'

This is because get_server_ip expects to be called with a detailed
server response, but in these cases is called with a create server
response, which does not contain the 'addresses' field.

This change fixes the issue by obtaining a detailed server response
where necessary.

Change-Id: I9deeb0e527a8537c71657507cabd81a63a421d9c
Closes-Bug: #1817701