Fix the logging of request body credentials
Recently we moved the logging of request from rest_client's
request() to raw_request which end up logging the credentials
also from Token client request.

This is to add a kwargs in raw_request() method which tell whether
request body needs to log or not. Also update Token clients to
pass '<omitted>' to stop logging the credential.

Change-Id: I7cbd3656463a51d18b4d72f45a324145283fc7eb