tempest cleanup - use admin_mgr only
Avoid using temporary managers and use only the admin one. The tool
has been using the admin_mgr for listing projects, roles, for listing
all resources during initializing a saved state so why don't use it also
for deleting all of the resources?
This wil solve the Unauthorized issue happening when the tool
was requesting credentials for the project from a different
domain.

Closes-bug: #1766582

Change-Id: Ibb5599ce48712b94dbf591e4a30cf3906cb5cdde