Add release note for log_req_body kwarg
This commit updates the release note from change
I7cbd3656463a51d18b4d72f45a324145283fc7eb. That commit was missing a
release note for the extra kwarg and concentrated only on the security
fix. This commit addresses that oversight. In addition the security note
is removed from the release notes because it's not addressing a released
security issue, just a short lived one introduced during development of
the pending release.

Change-Id: Id2d34eb911435e5a1036d0494a7df43eda2ebf0a