Do not specify a host for live-migration for non homogeneous nodes
In production cloud with non homogeneous compute nodes, specifying
a host for live-migration may fail unexpectedly, e.g., if source host
is with CPU "E5-2699 v4" and the selected target host is with CPU
"E5-2670 v3", the live-migration will fail because of the downgrade
issue.

So this is to add config option can_migrate_between_any_hosts which
can be set to False so that in non homogeneous cloud it can select
destination host automatically.

Change-Id: I05e9c8360b86326b7f58b2c0600fcb420bc31482
Closes-Bug: #1851185