Fix test_attach_detach_volume test case to not rely on user/libvirt supplied disk name
This patch removes user provided and libvirt assigned
value as that was based on assumption rather count the
disk name before and after attach/detach based on output
generated using 'lsblk' command  in the running instance
with the help of new method list_disks in client and
existing util remote client for remote ssh.

we observer this test case failing when the instance boot
with config drive format "vfat" enabled with two volume
vda and vdb volume disk name, so whatever test case boot with
config drive as well as ephemeral disk  was not be able to
use this testcase.

Change-Id: Id3e537bc55089ca5c35fa1f973dd47f2b11147ae
Partial-Bug: #1852807