Add default RBAC personas to pre-provisioned creds
This change adds support for the nine default personas for the
pre-provisioned credentials provider, as well as a sample accounts.yaml
file. With this, service projects should be able to use either dynamic
or pre-provisioned credentials for testing new policy defaults that use
system, domain, or project scope and the admin, member or reader default
roles.

Change-Id: Ifda77fb8d875b18c4507da2c9d98f17bcdac2e21