Add a default network when creating a server
If a new server is created without defining a network, Nova will
retrieve the available networks for the corresponding project. In case
of having more than one network, Nova raises the exception
"NetworkAmbiguous", informing that in case of multiple networks found,
one must be selected in the call.

This error has been detected in "ServersNegativeTestJSON" test cases.

Change-Id: I71333a45982fc0b6b155d68e96640a9db3d41e4a
Closes-Bug: #1844568