Create default network for several compute tests
Chipping away at the referenced bug, this commit introduces a
class-level variable on BaseV2ComputeTest that triggers
setup_credentials to create a default network, and sets it for several
classes where the problem has been observed. The class variable is so we
can whack-a-mole with one line instead of four whenever this problem
pops up.

Change-Id: Ia34fe3b41261de02918713c2a948e6c56554bf6f
Partial-Bug: #1844568