compute: Add negative volume update multiattach test
The introduction of Icd7fcb87a09c35a13e4e14235feb30a289d22778 blocks
attempting to update or swap a volume that is currently attached to
multiple instances. This change adds a simple test to assert that a
BadRequest is raised when this is attempted directly against the compute
API.

Future work will also cover this failure when part of a wider volume
retype or migration flow.

Change-Id: I931a810901cf7847a8259e44a30b0150aad859f2