Drop python 3.5 support from Tempest
Tempest, its plugins and their dependencies in ussuri
cycle are python-requires>=py3.6 which makes distro not
having python 3.6 to do hack to install py3.6 etc.

We cannot do >py3.6 in basepython so one way is to drop the
py3.5 support and basepython to py3 so that it can choose
the available py version in env.

It time to drop the py3.5 from Tempest. Last supported version of
Tempest for py3.5 is 23.0.0.

oslo drop py2 so we need to remove the py2 unit test job which
is failing and blocking gate.

There is seoarate work to drop the py2 things from Tempest in
compplete set.

Change-Id: I872dbb57df46d1193c0d409bb00d56f372d295fe
Related-Bug: 1861308