Remove 'gabbi_tempest_path' from devstack-tempest job
We need to move 'gabbi_tempest_path' var to gabbi-tempest plugin jobs side
which is the only place it is being used. This var is specific to gabbi-tempest
not used for any other devstack-tempest jobs so defining it in devstack-tempest
base job is not the correct way. We might face issue like bug#1821072 again in
future which can end up blocking all gate jobs.

Below patch is moving this var to gabbi-tempest plugin.
- https://review.opendev.org/#/c/705339/

Change-Id: Id96e9b3f1271fce50aae982fdf12871693107e1d
Closes-Bug: #1861592