compute: Use valid extra specs to test 'os-extra_specs' API
In Ussuri, nova is hoping to implement support for extra spec
validation. If or when this is completed, only extra specs that are
recognised will be permitted by default. We're currently using a number
of arbitrary extra specs to validate the 'os-extra_specs' API meaning we
will fall afoul of this validation. Head the issue off now by using
recognized extra specs, namely 'hw:numa_nodes' and 'hw:cpu_policy',
instead.

Change-Id: I95a1050d6a5e2e40793ae7f18ff0d326455b4b36
Signed-off-by: Stephen Finucane <sfinucan@redhat.com>