Adding description for testcases - compute part1
When Tempest is used in customer site, often we are required to
provide a testcase list including testcase names and descriptions.
Now no this kind of doc is available, however, we can extract info
from tempest code to create one, but, some testcases don't have
descriptions about what the testcases are testing, so, we can add
descrption for each testcase.

There are hundreds of testcases missing descriptions, so we can
add them gradually, and limit the modified files in one patch
for the convenience of reviewing.

Change-Id: Id90d6fedccc84aa9548fba62b5926a75050ab443
partially-implements: blueprint testcase-description