Ensure bootable volume is gone in scenario test before tearDown
In test_create_server_from_volume_snapshot tearDown would trigger
bootable volume and snapshot deletion in the correct order but not
wait for the bootable volume to be deleted, causing snapshot
deletion to fail.

This change makes the scenario test explicitly delete the bootable
volume to avoid the race condition in tearDown.

Change-Id: I3fe2ce349ce629bfc96f6ddfb45b22687a1be137
Co-Authored-By: Alan Bishop <abishop@redhat.com>
Closes-Bug: 1863750