Catch json.loads exceptions in verify_device_metadata
In case "TaggedBootDevicesTest.verify_device_metadata" "json.loads"
method returns an exception, catch it and return "False". This could
happen if, for example, the string to load is empty.

Because this function is called in an active wait loop, the content
string will be retrieved again and the check retried.

Change-Id: Id3a5b32f8ab197bcb1c59b5f6364dde4d8648eab
Closes-Bug: #1867904