Prevent raising exception if IP address is set in _hotplug_server
In "_hotplug_server", when the new interfaces is added to the server,
the function tries to configure the Neutron assigned IP address in
this new interface. Although the IP addresses are listed before to
check that the new port IP is not present, as reported in the bug,
sometimes the address is already set when the "ip a add" command is
executed.

Change-Id: I9594482735b482eb477f342dfd9f265a41eb002d
Closes-Bug: #1869899