Pass clients to compute base's create_test_server
Previously, create_test_server() in the base compute test class did
not provide a way to create a server with admin privileges. This can
be necessary when using APIs that are admin-only by default, for
example the `host` parameter when creating a server in Nova
microversion 2.74. This patch adds a clients kwarg to
create_test_server(). This kwarg is then passed to compute.common's
create_test_server() instead of the cls.os_primary clients that was
harcoded previously.

Change-Id: Ic47e6a4ced54739de9325bb14e332ba15cdb210f