Adding description for testcases - identity part1
When Tempest is used in customer site, often we are required to
provide a testcase list including testcase names and descriptions.
Now no this kind of doc is available, so we can add descriptions
with the format of doc string for every testcase, so later we
can generata such a testcase description list.

There are hundreds of testcases missing descriptions, so we can
add them gradually, and limit the modified files in one patch
for the convenience of reviewing.

BTW: some minor code refactor is done along with this patch.

Change-Id: I072bd2c86a7f50559e9e452dda3e1669fad3be00
partially-implements: blueprint testcase-description