Use stable constraint in tox to release new tag for Ussuri
We are going to releease Tempest new tag to declare the start of
support for stable/ussuri. So that new tag use stable/ussuri
constraint in the tox env.

For example: if anyone use Tempest 24.0.0 in future say 1 year later then
tox env also should use the ussuri contstraint for compatibility instead
master.

These need to be move back to master constraint once release is done.

By using master constraint in tox for all the old tag is problem and
we faced the issue of constraint incompatibility due to that and we
have to work around it all over the devstack, tempest role, grenade to
set the stable constraint cia env var.

- I4268049ff91e5527e4b95fd8a70fa53ba27d56ed

Change-Id: I39e3e23a85a9abc1eda6848a6f4af77540cb81b6