Move the grenade-py3 job to new zuulv3 grenade job name
'grenade' is zuulv3 native job for py3 (because devstack is default to
py3 since Ussuri) which replace the old 'grenade-py3'
legacy job. 'grenade-py3' job is alias to the new zuulv3 native
'grenade' job but that is only for migration time. Later once
all switch to new job name, we will remvoe the 'grenade-py3' alias
job.

Change-Id: Iea355514ec7cc2a52ed5e88c356bcb1cdeadd639