Move ceph job to voting
Currently ceph job is n-v, if I remember
correctly it was kept n-v due to its stability.

But it is stable and voting on many projects. With
n-v it is hard to monitor it all the time.

let's make it voting and if any new tests failing we
can add that in the ceph blacklist in advance.
Example: https://review.opendev.org/#/c/724866/

Change-Id: Ic484d233b325c887f0a19f3ec06376d403beb964