compute: Fix boot from volume stable rescue tests
The initial set of tests would incorrectly attempt to use a volume based
snapshot image as the rescue image. Such images contain no actual image
data and are just used to store metadata such as the
block_device_mapping that is in turn used when creating new instances
from the backing volume snapshots.

This wasn't initially noticed as the tests pass on the default file
based imagebackend in Nova that allows zero length sparse files to be
created. This failed however when the rbd imagebackend was used as the
import process into rbd does not accept zero sized files to be used.

To correct this an image based instance is now always created and used
to generate a suitable rescue image with each test. An additional
instance is then also created for the boot from volume based tests to
ensure the correct mapping is used.

The boot from volume snapshot test is also removed and the remaining
boot from volume tests are marked as slow given the additional overhead
of creating a second instance using these tests.

Change-Id: I7356b54bef0c614d9bfd1ed0d7b42574b58966f9
Closes-Bug: #1876330