Bump hacking min version to 3.0.1
hacking 3.0.1 fix the pinning of flake8 to avoid bringing in a new
version with new checks.

bumping the min version for hacking so that any older hacking versions
which auto adopt the new checks are not used.

The 2.6.0 version introduces some checks that cause failures
with the current code. hacking and flake8 cap could have solve
the pycodestyle compatible version but flake8-import-order which
does not cap the pycodestyle pulled its latest version and break
things.

Change-Id: I8996fec39e9b7c411b2708eb8baeb50aa2c28d42