Fix 404 when deleteing volume group
In delete_group() method the group is deleted first and
then the volume information is fetched which internally
trys to find all the related groups. As group delete request
was placed earlier than volume list and the group might be
deleted before volume list try to find the group
information, it returns 404.

This is race condition based on group is finish its deletion
before list volume requests or not.

Better is to fetch the information of volumes first and then
delete the group. And use that volume info later
to cleanup the volume also if requested.

Change-Id: I5b52ebac2f325bfeda51773c90cb40e7f7e02c67
Closes-Bug: #1882116