Add default headers to the volume show_version request
When openstack is deployed with kubernetes and openstack-helm
and when tempest test suite is run against that lab, 
test_versions.VersionsTest.test_show_version tempest test case
is failing. This issue is not reproducible on all the labs 
deployed with kubernetes and helm but only on one of the lab. 
There is nothing inconsistent found in the cinder functionality 
on that lab except this test case failure.

Adding 'Openstack-Api-Version' header to the "show_version" request 
as that of "list_versions" request solves the above issue.
'Openstack-Api-Version' header is included in default headers.
So, modifying "show_version" request to include default headers
as there will be no negative impacts.

Change-Id: I2f1970d4595c1b32c05f27c5c0eb3e28a75d7052

Closes-Bug: #1887789